feat: don't patch .desktop file on Linux
Some AppImages (compressed?) come without "[Desktop Entry]", so the
update failed while trying to read it. It was a ugly thing anyway, to
patch .desktop files to avoid the annoying "integrate with system?"
dialog; instead, we'll switch the AppImage release to silently integrate
with the system, so that we don't need to handle .desktop files in the
updater anymore.